refactor : 평균 소비 비교 컴포넌트 분리
